Refactored Notes input so it dynamically adjusts placeholder on one elment instead of 4
