+++ src/App.vue
@@ -22,56 +22,21 @@
 
 <script>
 	import jspdf from "jspdf";
+	import CustomerInformation from './components/customerInformation'
 	import EquipmentInput from "./components/equipment-input";
-	import CustomerInformation from "./components/customer-information";
 	import ReturnInformation from "./components/return-information"
 
 	export default {
 		components: {CustomerInformation, EquipmentInput, ReturnInformation},
 		data(){
 			return {
-				customerShit: [],
-				returnInformation: [],
-				equipmentInformation: [],
-
 				image: "https://www.hbci.com/wp-content/uploads/footer-logo.png",
 				logoAlt: "Hiawatha Broadband Communications",
 			};
 		},
 		methods: {
 			exportPDF() {
-				let heading = "Customer Information: \n";
-				let name = "Customer Name: " + this.firstName + " " + this.lastName + "\n";
-				let phone = "Contact Number: " + this.customerPhoneNumber + "\n";
-				let address = "Service Address: " + this.customerAddress + "\n";
-				let account = "Account #: " + this.customerAccount + "\n";
-
-				let equipmentType =
-					"Type of Equipment: " + document.EquipmentInput.equipmentType + "\n\n";
-				let equipmentNum = "CMAC/SN:" + this.equipmentNum + "\n";
-				let hasRemote = this.remote + "\n";
-				let hasPwrCord = this.powerCord + "\n";
-
-				let returnType = "Type of Return: " + this.returnType + "\n";
-				let notes = "Notes: " + this.notes + "\n";
-
-				let customerData = [heading, name, phone, address, account];
-				let equipmentData = [equipmentType, equipmentNum, hasRemote, hasPwrCord];
-				let returnData = [returnType, notes];
-
-				customerData = customerData.join();
-				equipmentData = equipmentData.join();
-				returnData = returnData.join();
-
-				console.log(customerData);
-				console.log(equipmentData);
-				console.log(returnData);
-
-				let data = [customerData, equipmentData, returnData];
-
-				data = data.join("\n");
-				console.log(data);
-
+				
 				let doc = new jspdf({
 					orientation: "p",
 					unit: "in",
@@ -81,7 +46,7 @@
 				doc.save("return.pdf");
 			},
 			printInputs(){
-				console.log(CustomerInformation.data().firstName);
+				console.log(CustomerInformation.firstName);
 			}
 		}
 	};
